Refactor create and import account
