--- src/redux/constants.js
@@ -1,38 +1,0 @@
-export const DEFAULT_ASSETS = [
-  {
-    symbol: 'ICP',
-    name: 'ICP',
-    amount: 0,
-    value: 0,
-    icon: 'dfinity',
-  },
-  {
-    symbol: 'XTC',
-    name: 'Cycles',
-    amount: 0,
-    value: 0,
-    icon: 'xtc',
-  },
-];
-
-export const TRANSACTION_STATUS = {
-  success: 'success',
-  error: 'error',
-};
-
-export const recursiveParseBigint = obj =>
-  Object.entries(obj).reduce(
-    (acum, [key, val]) => {
-      if (val instanceof Object) {
-        const res = Array.isArray(val)
-          ? val.map(el => recursiveParseBigint(el))
-          : recursiveParseBigint(val);
-        return { ...acum, [key]: res };
-      }
-      if (typeof val === 'bigint') {
-        return { ...acum, [key]: parseInt(val.toString(), 10) };
-      }
-      return { ...acum, [key]: val };
-    },
-    { ...obj },
-  );
