+++ src/screens/Send/components/ReviewSend/index.js
@@ -7,10 +7,10 @@ import RainbowButton from '../../../../components/buttons/RainbowButton';
 import NftDisplayer from '../../../../components/common/NftDisplayer';
 import TokenFormat from '../../../../components/number/TokenFormat';
 import { Colors, FontStyles } from '../../../../constants/theme';
-import { TRANSACTION_STATUS } from '../../../../redux/constants';
 import TokenIcon from '../../../../components/tokens/TokenIcon';
 import { setTransaction } from '../../../../redux/slices/user';
 import UserIcon from '../../../../components/common/UserIcon';
+import { TRANSACTION_STATUS } from '../../../../redux/utils';
 import shortAddress from '../../../../helpers/short-address';
 import Button from '../../../../components/buttons/Button';
 import Header from '../../../../components/common/Header';
